added wineinfo
